+++ CHANGELOG.md
@@ -10,6 +10,8 @@ All notable changes to this project will be documented in this file.
 
 - Converted DPR Interpolation script to Polars and added tests for the same.
 
+- Converted DPR Extrapolation script to Polars and added tests for the same.
+
 ### Changed
 - Removed the PySpark version of IND CQC Clean and Validation jobs.